requirements.txt
groq>=0.11.0
pyyaml>=6.0.0
mcp>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typer>=0.12.0
python-dotenv>=1.0.0
httpx>=0.27.0
anyio>=4.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
